+++ WebApp/Controllers/CenterController.cs
@@ -16,21 +16,21 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<List<GetCenterDto>>>> GetCenters()
     {
         var response = await centerService.GetCenters();
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("{id}")]
     public async Task<ActionResult<Response<GetCenterDto>>> GetCenterById(int id)
     {
         var response = await centerService.GetCenterByIdAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("paginated")]
     public async Task<ActionResult<PaginationResponse<List<GetCenterDto>>>> GetCentersPaginated([FromQuery] CenterFilter filter)
     {
         var response = await centerService.GetCentersPaginated(filter);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
     [HttpPost]
@@ -38,7 +38,7 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<string>>> CreateCenter([FromForm] CreateCenterDto createCenterDto)
     {
         var response = await centerService.CreateCenterAsync(createCenterDto);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
 
@@ -47,7 +47,7 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<string>>> UpdateCenter(int id, [FromForm] UpdateCenterDto updateCenterDto)
     {
         var response = await centerService.UpdateCenterAsync(id, updateCenterDto);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
 
@@ -56,14 +56,14 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<string>>> DeleteCenter(int id)
     {
         var response = await centerService.DeleteCenterAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
     [HttpGet("{id}/groups")]
     public async Task<ActionResult<Response<List<GetCenterGroupsDto>>>> GetCenterGroups(int id)
     {
         var response = await centerService.GetCenterGroupsAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
 
@@ -71,21 +71,21 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<List<GetCenterStudentsDto>>>> GetCenterStudents(int id)
     {
         var response = await centerService.GetCenterStudentsAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
     [HttpGet("{id}/mentors")]
     public async Task<ActionResult<Response<List<GetCenterMentorsDto>>>> GetCenterMentors(int id)
     {
         var response = await centerService.GetCenterMentorsAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
     [HttpGet("{id}/courses")]
     public async Task<ActionResult<Response<List<GetCenterCoursesDto>>>> GetCenterCourses(int id)
     {
         var response = await centerService.GetCenterCoursesAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
 
@@ -93,6 +93,6 @@ public class CenterController(ICenterService centerService) : ControllerBase
     public async Task<ActionResult<Response<CenterStatisticsDto>>> GetCenterStatistics(int id)
     {
         var response = await centerService.GetCenterStatisticsAsync(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 }
