+++ WebApp/Controllers/CommentController.cs
@@ -15,42 +15,42 @@ public class CommentController(ICommentService commentService) : ControllerBase
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetAllComments()
     {
         var response = await commentService.GetComments();
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 
     [HttpGet("{id}")]
     public async Task<ActionResult<Response<GetCommentDto>>> GetCommentById(int id)
     {
         var response = await commentService.GetCommentById(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("student/{studentId}")]
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetCommentsByStudent(int studentId)
     {
         var response = await commentService.GetCommentsByStudent(studentId);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("group/{groupId}")]
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetCommentsByGroup(int groupId)
     {
         var response = await commentService.GetCommentsByGroup(groupId);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("lesson/{lessonId}")]
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetCommentsByLesson(int lessonId)
     {
         var response = await commentService.GetCommentsByLesson(lessonId);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("type/{type}")]
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetCommentsByType(CommentType type)
     {
         var response = await commentService.GetCommentsByType(type);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpGet("private/{authorId}")]
@@ -58,7 +58,7 @@ public class CommentController(ICommentService commentService) : ControllerBase
     public async Task<ActionResult<Response<List<GetCommentDto>>>> GetPrivateComments(int authorId)
     {
         var response = await commentService.GetPrivateComments(authorId);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpPost]
@@ -66,7 +66,7 @@ public class CommentController(ICommentService commentService) : ControllerBase
     public async Task<ActionResult<Response<string>>> CreateComment([FromBody] CreateCommentDto createCommentDto)
     {
         var response = await commentService.CreateComment(createCommentDto);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpPut]
@@ -74,7 +74,7 @@ public class CommentController(ICommentService commentService) : ControllerBase
     public async Task<ActionResult<Response<string>>> UpdateComment([FromBody] UpdateCommentDto updateCommentDto)
     {
         var response = await commentService.UpdateComment(updateCommentDto);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
     
     [HttpDelete("{id}")]
@@ -82,6 +82,6 @@ public class CommentController(ICommentService commentService) : ControllerBase
     public async Task<ActionResult<Response<string>>> DeleteComment(int id)
     {
         var response = await commentService.DeleteComment(id);
-        return StatusCode((int)response.StatusCode, response);
+        return StatusCode(response.StatusCode, response);
     }
 }
