+++ WebApp/Controllers/UserController.cs
@@ -17,7 +17,7 @@ public class UserController(IUserService service) : ControllerBase
     public async Task<ActionResult<PaginationResponse<List<GetUserDto>>>> GetAllUsers([FromQuery] UserFilter filter)
     {
         var result = await service.GetUsersAsync(filter);
-        return StatusCode((int)result.StatusCode, result);
+        return StatusCode(result.StatusCode, result);
     }
 
     [HttpGet("{id}")]
@@ -25,7 +25,7 @@ public class UserController(IUserService service) : ControllerBase
     public async Task<ActionResult<Response<GetUserDto>>> GetUserById(int id)
     {
         var result = await service.GetUserByIdAsync(id);
-        return StatusCode((int)result.StatusCode, result);
+        return StatusCode(result.StatusCode, result);
     }
         
     [HttpGet("me")]
@@ -33,7 +33,7 @@ public class UserController(IUserService service) : ControllerBase
     public async Task<ActionResult<Response<GetUserDto>>> GetCurrentUser()
     {
         var result = await service.GetCurrentUserAsync();
-        return StatusCode((int)result.StatusCode, result);
+        return StatusCode(result.StatusCode, result);
     }
     
     [HttpGet("search")]
@@ -41,7 +41,7 @@ public class UserController(IUserService service) : ControllerBase
     public async Task<ActionResult<Response<List<GetUserDto>>>> SearchUsers([FromQuery] string searchTerm)
     {
         var result = await service.SearchUsersAsync(searchTerm);
-        return StatusCode((int)result.StatusCode, result);
+        return StatusCode(result.StatusCode, result);
     }
     
     [HttpGet("role/{role}")]
@@ -49,14 +49,6 @@ public class UserController(IUserService service) : ControllerBase
     public async Task<ActionResult<Response<List<GetUserDto>>>> GetUsersByRole(string role)
     {
         var result = await service.GetUsersByRoleAsync(role);
-        return StatusCode((int)result.StatusCode, result);
+        return StatusCode(result.StatusCode, result);
     }
-    //
-    // [HttpGet("{userId}/activity")]
-    // [Authorize(Roles = $"{Roles.Admin},{Roles.SuperAdmin}")]
-    // public async Task<ActionResult<Response<UserActivityDto>>> GetUserActivity(int userId)
-    // {
-    //     var result = await service.GetUserActivityAsync(userId);
-    //     return StatusCode((int)result.StatusCode, result);
-    // }
 }
